Add basic go stuff to build and run the project
Dependabot: We are making sure the project keeps their dependencies up to date.
GH Actions: We are using several actions some for the CI pipeline but some others to make sure dependabot PRs are merged automatically if they pass CI. Aditionally we are bringing [Shipit](https://www.kevinrobayna.com/2022/08/avoid_breaking_main_branch)!

Makefile: To ensure we are all building the app the same way.

CODE_OF_CONDUCT: To ensure we all understand how we are expected to behave.