Allow app to receive slash commands (#18)
This PR adds the required code to handle slash command requests from
slack. Currently the only command it's `/rotabot`.

Why there are no e2e tests?
This PR is not adding a test to ensure we parse the requests from slack
correctly because it's "quite a pain". We are currently using the
slack-go library which ensures the parsing happens correctly although we
should still test it. We will add such tests in the future once we have
some functionality as we would need to either mock the request signature
verifier or copy some code from the slack-go library to generate an
unsafe request i.e
https://github.com/slack-go/slack/blob/c99b47be75f6f1cc18709295540336de85743358/security.go#L4